// src/app/libraries/lily/layout/div/models/enums/horizontal-align.ts
/**
 * Describes the horizontal alignment of elements within a lily div.
 */
export enum ELilyDivHorizontalAlign {
  /**
   * Aligns the elements to the left of the div.
   */
  Left = 'lily-div-horizontal-align-left',

  /**
   * Aligns the elements to the center of the div.
   */
  Center = 'lily-div-horizontal-align-center',

  /**
   * Aligns the elements to the right of the div.
   */
  Right = 'lily-div-horizontal-align-right',
}