// src/app/libraries/lily/layout/div/models/enums/vertical-align.ts
/**
 * Describes the vertical alignment of elements wiithin a lily div.
 */
export enum ELilyDivVerticalAlign {
  /**
   * Aligns the elements to the top of the div.
   */
  Top = 'lily-div-vertical-align-top',

  /**
   * Aligns the elements to the center of the div.
   */
  Center = 'lily-div-vertical-align-center',

  /**
   * Aligns the elements to the bottom of the div.
   */
  Bottom = 'lily-div-vertical-align-bottom',
}